refactor(framework-core): extract shared TempDir helper into test_utils
idl_gen.rs and extension.rs duplicated the same TempDir + Drop + write
helper. Move to crate-internal `spel-framework-core::test_utils`
(cfg(test)) and reuse.
