+++ spel-framework-core/src/extension.rs
@@ -143,30 +143,7 @@ pub fn discover_extension_instruction_attrs(
 mod tests {
     use super::*;
     use crate::idl::{IdlSeed, IdlType, SpelIdl};
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    static COUNTER: AtomicUsize = AtomicUsize::new(0);
-
-    struct TempDir(PathBuf);
-
-    impl TempDir {
-        fn new(label: &str) -> Self {
-            let n = COUNTER.fetch_add(1, Ordering::Relaxed);
-            let path = std::env::temp_dir().join(format!("spel-ext-test-{label}-{n}"));
-            std::fs::create_dir_all(&path).unwrap();
-            TempDir(path)
-        }
-        fn write(&self, rel: &str, content: &str) -> PathBuf {
-            let p = self.0.join(rel);
-            std::fs::create_dir_all(p.parent().unwrap()).unwrap();
-            std::fs::write(&p, content).unwrap();
-            p
-        }
-        fn path(&self) -> &Path { &self.0 }
-    }
-
-    impl Drop for TempDir {
-        fn drop(&mut self) { std::fs::remove_dir_all(&self.0).ok(); }
-    }
+    use crate::test_utils::TempDir;
 
     #[test]
     fn read_spel_extension_attr_returns_declared_name() {
