+++ spel-framework-core/src/test_utils.rs
@@ -1,0 +1,27 @@
+#[cfg(test)]
+
+use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicUsize, Ordering};
+static COUNTER: AtomicUsize = AtomicUsize::new(0);
+
+pub(crate) struct TempDir(PathBuf);
+
+impl TempDir {
+    pub fn new(label: &str) -> Self {
+        let n = COUNTER.fetch_add(1, Ordering::Relaxed);
+        let path = std::env::temp_dir().join(format!("spel-ext-test-{label}-{n}"));
+        std::fs::create_dir_all(&path).unwrap();
+        TempDir(path)
+    }
+    pub fn write(&self, rel: &str, content: &str) -> PathBuf {
+        let p = self.0.join(rel);
+        std::fs::create_dir_all(p.parent().unwrap()).unwrap();
+        std::fs::write(&p, content).unwrap();
+        p
+    }
+    pub fn path(&self) -> &Path { &self.0 }
+}
+
+impl Drop for TempDir {
+    fn drop(&mut self) { std::fs::remove_dir_all(&self.0).ok(); }
+}
